+++ web/app/error.tsx
@@ -1,0 +1,16 @@
+'use client'
+import Paper from '@mui/material/Paper'
+import Typography from '@mui/material/Typography'
+import Link from 'next/link'
+import ListIcon from '@mui/icons-material/List'
+import Fab from '@mui/material/Fab'
+
+export default function Error({error}: {error?: Error}) {
+  return (
+    <Paper sx={{ width: '100%', textAlign: 'center', padding: 2, minHeight: 100 }}>
+      <Fab sx={{ float: 'left', marginLeft: 1, marginTop: 1 }} size="small" color="primary" component={Link} href="/"><ListIcon /></Fab>
+      <Typography variant="h6">Error</Typography>
+      <Typography variant="body1">{error?.message}</Typography>
+    </Paper>
+  )
+}
