+++ web/app/page.tsx
@@ -1,6 +1,6 @@
 import SearchBox from "@/lib/components/search-box";
 import type { Metadata } from 'next'
-
+import Searcher from "@/lib/utils/searcher";
 const description = process.env.SITE_DESCRIPTION || ''
 
 export const metadata: Metadata = {
@@ -14,6 +14,9 @@ export const metadata: Metadata = {
 
 export default function Page() {
   return (
-    <SearchBox />
+    <>
+      <Searcher />
+      <SearchBox />
+    </>
   )
 }
